requirements.txt
locust
pyyaml
pandas
seaborn
matplotlib
numpy
fastparquet
pytest
pyarrow
